requirements.txt
Django==1.8.6
apt-xapian-index==0.45
argparse==1.2.1
base58==0.2.2
bitcoin==1.1.36
bitmerchant==0.1.8
blocktrail-sdk==1.0.6
cachetools==1.1.5
chardet==2.0.1
colorama==0.2.5
django-dia==0.2
django-easy-pjax==1.2.0
django-material==0.4.1
django-tastypie==0.12.2
django-tastypie-crust==0.2
ecdsa==0.13
future==0.15.2
html5lib==0.999
httpsig==1.1.2
psycopg2==2.6.1
pycrypto==2.6.1
python-apt==0.9.3.5
python-dateutil==2.4.2
python-debian==0.1.21-nmu2ubuntu2
python-mimeparse==0.1.4
requests==2.9.1
six==1.10.0
ssh-import-id==3.21
urllib3==1.7.1
wheel==0.24.0
wsgiref==0.1.2
